+++ packages/devtools_app/lib/src/screens/performance/tabbed_performance_view.dart
@@ -1,0 +1,171 @@
+// Copyright 2022 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+import 'package:flutter/material.dart';
+
+import '../../analytics/constants.dart' as analytics_constants;
+import '../../charts/flame_chart.dart';
+import '../../primitives/auto_dispose_mixin.dart';
+import '../../shared/common_widgets.dart';
+import '../../shared/theme.dart';
+import '../../ui/search.dart';
+import '../../ui/tab.dart';
+import 'frame_analysis.dart';
+import 'performance_controller.dart';
+import 'performance_model.dart';
+import 'performance_screen.dart';
+import 'raster_metrics.dart';
+import 'timeline_flame_chart.dart';
+
+final timelineSearchFieldKey = GlobalKey(debugLabel: 'TimelineSearchFieldKey');
+
+class TabbedPerformanceView extends StatefulWidget {
+  const TabbedPerformanceView({
+    required this.controller,
+    required this.processing,
+    required this.processingProgress,
+  });
+
+  final PerformanceController controller;
+
+  final bool processing;
+
+  final double processingProgress;
+
+  @override
+  _TabbedPerformanceViewState createState() => _TabbedPerformanceViewState();
+}
+
+class _TabbedPerformanceViewState extends State<TabbedPerformanceView>
+    with AutoDisposeMixin, SearchFieldMixin<TabbedPerformanceView> {
+  static const _gaPrefix = 'performanceTab';
+
+  PerformanceController get controller => widget.controller;
+
+  FlutterFrame? _selectedFlutterFrame;
+
+  @override
+  void initState() {
+    super.initState();
+
+    _selectedFlutterFrame = controller.selectedFrame.value;
+    addAutoDisposeListener(controller.selectedFrame, () {
+      setState(() {
+        _selectedFlutterFrame = controller.selectedFrame.value;
+      });
+    });
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    late Widget frameAnalysisView;
+    final selectedFrame = _selectedFlutterFrame;
+    if (selectedFrame != null) {
+      frameAnalysisView = FlutterFrameAnalysisView(
+        frameAnalysis: selectedFrame.frameAnalysis,
+      );
+    } else {
+      frameAnalysisView = const Center(
+        child: Text('Select a frame above to view analysis data.'),
+      );
+    }
+
+    final rasterMetrics = Center(
+      child: RenderingLayerVisualizer(
+        rasterMetricsController: controller.rasterMetricsController,
+      ),
+    );
+
+    final tabViews = [
+      TimelineEventsView(
+        controller: controller,
+        processing: widget.processing,
+        processingProgress: widget.processingProgress,
+      ),
+      if (frameAnalysisSupported) frameAnalysisView,
+      if (rasterMetricsSupported) rasterMetrics,
+    ];
+
+    return AnalyticsTabbedView(
+      tabs: _generateTabs(),
+      tabViews: tabViews,
+      gaScreen: analytics_constants.performance,
+      // TODO(kenz): enable analytics when this view is stable.
+      sendAnalytics: false,
+    );
+  }
+
+  List<DevToolsTab> _generateTabs() {
+    final data = controller.data;
+    final hasData = data != null && !data.isEmpty;
+    final searchFieldEnabled = hasData && !widget.processing;
+    return [
+      _buildTab(
+        tabName: 'Timeline Events',
+        trailing: Row(
+          mainAxisAlignment: MainAxisAlignment.end,
+          children: [
+            _buildSearchField(searchFieldEnabled),
+            const FlameChartHelpButton(
+              gaScreen: PerformanceScreen.id,
+              gaSelection: analytics_constants.timelineFlameChartHelp,
+            ),
+            RefreshTimelineEventsButton(controller: controller),
+          ],
+        ),
+      ),
+      if (frameAnalysisSupported)
+        _buildTab(
+          tabName: 'Frame Analysis',
+        ),
+      if (rasterMetricsSupported)
+        _buildTab(
+          tabName: 'Raster Metrics',
+          trailing: Row(
+            mainAxisAlignment: MainAxisAlignment.end,
+            children: [
+              IconLabelButton(
+                tooltip:
+                    'Take a snapshot of the rendering layers on the current'
+                    ' screen',
+                icon: Icons.camera,
+                label: 'Take Snapshot',
+                outlined: false,
+                onPressed:
+                    controller.rasterMetricsController.collectRasterStats,
+              ),
+              const SizedBox(width: denseSpacing),
+              ClearButton(
+                outlined: false,
+                onPressed: controller.rasterMetricsController.clear,
+              ),
+              const SizedBox(width: densePadding),
+            ],
+          ),
+        ),
+    ];
+  }
+
+  Widget _buildSearchField(bool searchFieldEnabled) {
+    return Container(
+      width: defaultSearchTextWidth,
+      height: defaultTextFieldHeight,
+      child: buildSearchField(
+        controller: controller,
+        searchFieldKey: timelineSearchFieldKey,
+        searchFieldEnabled: searchFieldEnabled,
+        shouldRequestFocus: false,
+        supportsNavigation: true,
+      ),
+    );
+  }
+
+  DevToolsTab _buildTab({required String tabName, Widget? trailing}) {
+    return DevToolsTab.create(
+      tabName: tabName,
+      gaPrefix: _gaPrefix,
+      trailing: trailing,
+    );
+  }
+}
