+++ packages/devtools_app/test/tabbed_performance_view_test.dart
@@ -1,18 +1,17 @@
-// Copyright 2019 The Chromium Authors. All rights reserved.
+// Copyright 2022 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:io';
 
 import 'package:devtools_app/src/charts/flame_chart.dart';
 import 'package:devtools_app/src/config_specific/ide_theme/ide_theme.dart';
 import 'package:devtools_app/src/config_specific/import_export/import_export.dart';
+import 'package:devtools_app/src/screens/performance/frame_analysis.dart';
 import 'package:devtools_app/src/screens/performance/performance_controller.dart';
 import 'package:devtools_app/src/screens/performance/performance_model.dart';
-import 'package:devtools_app/src/screens/performance/performance_screen.dart';
 import 'package:devtools_app/src/screens/performance/raster_metrics.dart';
-import 'package:devtools_app/src/screens/performance/timeline_analysis.dart';
+import 'package:devtools_app/src/screens/performance/tabbed_performance_view.dart';
 import 'package:devtools_app/src/screens/performance/timeline_flame_chart.dart';
 import 'package:devtools_app/src/service/service_manager.dart';
 import 'package:devtools_app/src/shared/common_widgets.dart';
@@ -21,7 +20,6 @@ import 'package:devtools_app/src/shared/preferences.dart';
 import 'package:devtools_app/src/shared/version.dart';
 import 'package:devtools_app/src/ui/tab.dart';
 import 'package:devtools_test/devtools_test.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/mockito.dart';
@@ -191,117 +189,5 @@ void main() {
         expect(find.byType(ClearButton), findsOneWidget);
       });
     });
-  });
-
-  group('TimelineAnalysisContainer', () {
-    setUp(() async {
-      await _setUpServiceManagerWithTimeline(testTimelineJson);
-    });
-
-    Future<void> pumpPerformanceScreenBody(
-      WidgetTester tester, {
-      PerformanceController? performanceController,
-      bool runAsync = false,
-    }) async {
-      controller = performanceController ?? PerformanceController();
-
-      if (runAsync) {
-        // Await a small delay to allow the PerformanceController to complete
-        // initialization.
-        await Future.delayed(const Duration(seconds: 1));
-      }
-
-      await tester.pumpWidget(
-        wrapWithControllers(
-          const PerformanceScreenBody(),
-          performance: controller,
-        ),
-      );
-      await tester.pumpAndSettle();
-    }
-
-    const windowSize = Size(2225.0, 1000.0);
-
-    testWidgetsWithWindowSize('builds header with search field', windowSize,
-        (WidgetTester tester) async {
-      await tester.runAsync(() async {
-        await _setUpServiceManagerWithTimeline({});
-        await pumpPerformanceScreenBody(tester);
-        await tester.pumpAndSettle();
-        expect(find.text('Timeline Events'), findsOneWidget);
-        expect(find.byType(RefreshTimelineEventsButton), findsOneWidget);
-        expect(find.byKey(timelineSearchFieldKey), findsOneWidget);
-        expect(find.byType(FlameChartHelpButton), findsOneWidget);
-      });
-    });
-
-    testWidgetsWithWindowSize('can show help dialog', windowSize,
-        (WidgetTester tester) async {
-      await tester.runAsync(() async {
-        await _setUpServiceManagerWithTimeline({});
-        await pumpPerformanceScreenBody(tester);
-        await tester.pumpAndSettle();
-
-        final helpButtonFinder = find.byType(FlameChartHelpButton);
-        expect(helpButtonFinder, findsOneWidget);
-        await tester.tap(helpButtonFinder);
-        await tester.pumpAndSettle();
-        expect(find.text('Flame Chart Help'), findsOneWidget);
-      });
-    });
-
-    testWidgetsWithWindowSize('builds flame chart with data', windowSize,
-        (WidgetTester tester) async {
-      await tester.runAsync(() async {
-        await pumpPerformanceScreenBody(tester, runAsync: true);
-        expect(find.byType(TimelineFlameChart), findsOneWidget);
-        expect(
-          find.byKey(TimelineEventsView.emptyTimelineKey),
-          findsNothing,
-        );
-      });
-    });
-
-    testWidgetsWithWindowSize('builds flame chart with no data', windowSize,
-        (WidgetTester tester) async {
-      await tester.runAsync(() async {
-        await _setUpServiceManagerWithTimeline({});
-        await pumpPerformanceScreenBody(tester, runAsync: true);
-        await tester.pumpAndSettle();
-        expect(find.byType(TimelineFlameChart), findsNothing);
-        expect(
-          find.byKey(TimelineEventsView.emptyTimelineKey),
-          findsOneWidget,
-        );
-      });
-    });
-
-    testWidgetsWithWindowSize(
-      'builds flame chart with selected frame',
-      windowSize,
-      (WidgetTester tester) async {
-        await tester.runAsync(() async {
-          await pumpPerformanceScreenBody(tester, runAsync: true);
-          controller
-            ..addFrame(testFrame1.shallowCopy())
-            ..addTimelineEvent(goldenUiTimelineEvent)
-            ..addTimelineEvent(goldenRasterTimelineEvent);
-          final data = controller.data!;
-          expect(data.frames.length, equals(1));
-          await controller.toggleSelectedFrame(data.frames.first);
-          await tester.pumpAndSettle();
-        });
-        expect(find.byType(TimelineFlameChart), findsOneWidget);
-        await expectLater(
-          find.byType(TimelineFlameChart),
-          matchesGoldenFile(
-            'goldens/timeline_flame_chart_with_selected_frame.png',
-          ),
-        );
-        // Await delay for golden comparison.
-        await tester.pumpAndSettle(const Duration(seconds: 2));
-      },
-      skip: kIsWeb || !Platform.isMacOS,
-    );
   });
 }
